Admin console should show available devices
To simplify adding new device mappings, the admin console is supposed to show the devices that are available on the configured servers. This functionality is broken in the current state of the code combined with the server that I'm testing with.

The cause seems to be a change in format of the result of the SIP show peers command.

This change should allow the code to process both the old as well as the new format (both formats are described in comments in the code that is added in this commit).

fixes #10